ml_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.20
pillow==10.4.0
torch==2.8.0
numpy==2.1.2
opencv-python-headless==4.10.0.84
torchvision==0.21.0


